Move MuxPeer to Ouroboros.Network.Mux module
This is first patch towards makeing `network-mux` independent of
`typed-protcols`.   The `Ouroboros.Network.Mux` can expose everything we
need from `netweork-mux`.
