+++ ouroboros-network/ouroboros-network.cabal
@@ -48,6 +48,7 @@ library
                        Ouroboros.Network.BlockFetch.State
                        Ouroboros.Network.ChainFragment
                        Ouroboros.Network.DeltaQ
+                       Ouroboros.Network.Mux
                        Ouroboros.Network.NodeToNode
                        Ouroboros.Network.NodeToClient
                        Ouroboros.Network.Point
@@ -159,6 +160,7 @@ test-suite tests
                        Ouroboros.Network.ChainFragment
                        Ouroboros.Network.ChainProducerState
                        Ouroboros.Network.DeltaQ
+                       Ouroboros.Network.Mux
                        Ouroboros.Network.Node
                        Ouroboros.Network.NodeToNode
                        Ouroboros.Network.Point
